requirements.txt
google-genai
dotenv
tqdm
rouge-score